feat: implements `session_request_sent` handling in universal-provider
