+++ providers/universal-provider/src/providers/eip155.ts
@@ -1,7 +1,7 @@
 import Client from "@walletconnect/sign-client";
 import { JsonRpcProvider } from "@walletconnect/jsonrpc-provider";
 import { HttpConnection } from "@walletconnect/jsonrpc-http-connection";
-import { EngineTypes, SessionTypes } from "@walletconnect/types";
+import { EngineTypes, SessionTypes, SignClientTypes } from "@walletconnect/types";
 
 import {
   IProvider,
@@ -11,9 +11,8 @@ import {
   SessionNamespace,
 } from "../types";
 
-import { getRpcUrl, deeplinkRedirect } from "../utils";
+import { getRpcUrl, deeplinkRedirect, handleDeepLinks } from "../utils";
 import EventEmitter from "events";
-import { RELAYER_EVENTS } from "@walletconnect/core";
 
 class Eip155Provider implements IProvider {
   public name = "eip155";
@@ -49,7 +48,7 @@ class Eip155Provider implements IProvider {
         break;
     }
     if (this.namespace.methods.includes(args.request.method)) {
-      this.client.core.relayer.once(RELAYER_EVENTS.publish, deeplinkRedirect);
+      handleDeepLinks(this.client, args);
       return await this.client.request(args as EngineTypes.RequestParams);
     }
     return this.getHttpProvider().request(args.request);
