+++ providers/universal-provider/src/providers/solana.ts
@@ -10,7 +10,7 @@ import {
   SessionNamespace,
   SubProviderOpts,
 } from "../types";
-import { getRpcUrl, deeplinkRedirect } from "../utils";
+import { getRpcUrl, deeplinkRedirect, handleDeepLinks } from "../utils";
 import { RELAYER_EVENTS } from "@walletconnect/core";
 
 class SolanaProvider implements IProvider {
@@ -71,7 +71,7 @@ class SolanaProvider implements IProvider {
 
   public request<T = unknown>(args: RequestParams): Promise<T> {
     if (this.namespace.methods.includes(args.request.method)) {
-      this.client.core.relayer.once(RELAYER_EVENTS.publish, deeplinkRedirect);
+      handleDeepLinks(this.client, args);
       return this.client.request(args as EngineTypes.RequestParams);
     }
     return this.getHttpProvider().request(args.request);
