+++ providers/universal-provider/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./misc";
+export * from "./deepLinks";
